Fix homepage routing
