+++ client/src/pages/CreateEvaluation/EvaluationSummary/EvaluationSummary.jsx
@@ -127,7 +127,7 @@ function EvaluationSummary() {
         title='Kutsut lähetetty!'
         body='Lorem ipsum, dolor sit amet consectetur adipisicing elit'
         open={openNotificationModal}
-        redirectLink='/userdashboard'
+        redirectLink='/customer-list'
       />
     </main>
   );
