+++ client/src/pages/ForgotPassword/_forgotpassword.scss
@@ -1,23 +1,20 @@
 .app__wrapper {
-  // main wrapper for the forgot password page
+  
   .forgotPassword__wrapper {
     display: flex;
     flex-direction: column;
     justify-content: center;
-    align-items: center;
-    height: 100%;
-    h1{
+    height: 100vh;
+    width: 100%;
+    text-align: center;
+
+    h1 {
       color: var(--saukko-main-white);
     }
 
     .forgotPassword__container {
-      display: flex;
-      flex-direction: column;
-      padding: 0px 36px 0px 36px;
-      margin-bottom: 150px;
-      width: 100%;
-      height:15%;
-      gap: 10px;
+      padding: 30px;
+      margin-top: 80px;
 
       h2 {
         font-size: 24px;
