+++ client/src/pages/Home/UnitList/_unitlist.scss
@@ -1,0 +1,18 @@
+.unitList__wrapper {
+
+  .unitList__notifications {
+    margin-top: 55%;
+  
+  }
+  
+  .unitList__units {
+    margin-bottom: 80px;
+  }
+  
+  h3 {
+    margin-left: 15px;
+    margin-bottom: 20px;
+    font-size: 20px;
+  }
+}
+
