+++ client/src/pages/LandingPage/_landingpage.scss
@@ -8,16 +8,6 @@
     width: 100%;
     padding: 36px;
     background: var(--saukko-main-blue);
-    text-align: center;.app__wrapper {
-  
-  .landingPage__wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100vh;
-    width: 100%;
-    padding: 36px;
-    background: var(--saukko-main-blue);
     text-align: center;
 
     h1 {
@@ -41,28 +31,3 @@
     }
   }
 }
-
-    h1 {
-      font-size: 78px;
-      font-weight: bold;
-      color: var(--saukko-main-white);
-    }
-
-    p {
-      font-size: 18px;
-      color: var(--saukko-main-white);
-    }
-
-    .landingPage__textContainer {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-      gap: 20px;
-    }
-  }
-}
-
-
-
