--- client/src/pages/UserDashboard/_userdashboard.scss
@@ -1,14 +1,0 @@
-
-.app__wrapper{
-  .dashboardPage__wrapper{
-    display: flex;
-    flex-direction: column;
-
-   
-  }
-}
-.headingStyle{
-  font-size: 20px;
-  margin-left: 15px;
-  margin-bottom: 10px;
-}
