+++ client/src/scss/index.scss
@@ -32,8 +32,8 @@
 @import '../pages/AdminDegree/ConfirmSelection/confirmselection';
 @import '../pages/AdminDegree/SpecifyTasks/specifytasks';
 @import '../pages/AdminDegree/Summary/summary';
-@import '../pages/UserDashboard/userdashboard';
-@import '../pages/UserDashboard/CustomerList/customerlist';
+@import '../pages/Home/CustomerList/customerlist';
+@import '../pages/Home/UnitList/unitlist';
 @import '../pages/ContractInfo/contractinfo';
 @import '../pages/UpdateHomePageAfterLogin/updatehomepageafterlogin';
 @import '../pages/CompanyInfo/companyinfo';
